WP Slider
A client wanted a dynamic slider that they could easily add slides to
from the word press back-end. A neat way to do this is to create custom
taxonomy and custom post type. To make it easy for the client to add
slides we set up the loop so that it calls in the featured image as the
background image for the slide and the_content from the wysiwyg
displays on the slider…The title is not called and so the slides can
just be named Slide 1, Slide 2, or w/e. You will have to set up a
custom query. Basically we are creating a “sliding blog roll”